requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
openai>=1.12.0
mysql-connector-python>=8.3.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
